refactor(jssdk): signers
