+++ frontend/packages/sdk/src/index.ts
@@ -11,6 +11,7 @@ export * from './OnChainRegistry'
 export * from './options'
 export * from './abis/fetchers'
 export * from './pruntime/eip712'
+export * from './signers/WalletClientSigner'
 export { default as createPruntimeClient } from './pruntime/createPruntimeClient'
 export { default as signAndSend, SignAndSendError } from './utils/signAndSend'
 export { ackFirst } from './ha/ack-first'
